+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This is the log of notable changes to EAS CLI and related packages.
 
 - [eas-cli] Add `eas update:insights <groupId>` command to display launch, crash, unique-user, and payload-size metrics for an update group. ([@kadikraman](https://github.com/kadikraman))
 - [eas-cli] Add `--insights` flag to `eas update:view <groupId>` to append the same metrics below the existing output. ([@kadikraman](https://github.com/kadikraman))
+- [eas-cli] Add `eas channel:insights --channel <name> --runtime-version <version>` command to display adoption and crash metrics for a channel + runtime version. ([@kadikraman](https://github.com/kadikraman))
 
 ### 🐛 Bug fixes